pyproject.toml
[tool.poetry]
name = "pixelgrid"
version = "0.1.0"
description = ""
authors = ["Tarcísio Eduardo Moreira Crocomo <tarcisioe@pm.me>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.2"
Pillow = "^9.1.1"
typer = "^0.4.1"

[tool.poetry.dev-dependencies]
summon-python = {extras = ["core"], version = "^0.1.3"}
types-Pillow = "^9.0.19"
mypy = "^0.960"
flake8 = "^4.0.1"
pylint = "^2.14.0"
pytest-datadir = "^1.3.1"
pytest-cov = "^3.0.0"
pre-commit = "^2.19.0"
black = "^22.3.0"
isort = "^5.10.1"

[tool.summon.plugins.python]
test-modules = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
